add adaptive on collection page, add createPoolButton component
