+++ src/pages/Collection/components/CreatePoolButton/CreatePoolButton.module.scss
@@ -1,0 +1,3 @@
+.poolButton {
+  padding: 8px 16px;
+}
